Change en_i signal logic in clk_int_div for faster response time
Since the output clock gate is glitch free by design, it is perfectly fine to
drive the clk_gate enable signal with combinational logic. With this change, the
deasserting the en_i signal disables the clock immediately (next falling edge)
in contrast to the previous implementation which had 1 cycle delay.